2engine/app/requirements.txt
celery==5.4.0
gevent==24.2.1
flower==2.0.1

Django==5.0.7
django-elasticsearch-dsl==8.0
djangorestframework==3.15.2

psycopg2-binary==2.9.9

gunicorn==19.7.1